+++ 10OperacoesMatematicas.cs
@@ -1,32 +1,33 @@
 using System;
 
-namespace ExerciciosDois
+namespace ListaExercicios2
 {
-    public class OperacoesMatematicas 
+    public class OperacoesMatematicas
     {
         public static void Operacoes()
         {
-        int valorA = 0;
-        int valorB = 0;
-        int adicao;
-        int subtracao;
-        int multiplicacao;
-        int divisao;
-        int menu = 0;
+            int valorA = 0;
+            int valorB = 0;
+            int adicao;
+            int subtracao;
+            int multiplicacao;
+            int divisao;
+            int menu = 0;
 
-        Console.WriteLine("\nInforme o 1º Valor: ");
-        valorA = Int32.Parse(Console.ReadLine());
-        Console.WriteLine("Informe o 2º Valor: ");
-        valorB = Int32.Parse(Console.ReadLine());
+            Console.WriteLine("\nInforme o 1º Valor: ");
+            valorA = Int32.Parse(Console.ReadLine());
+            Console.WriteLine("Informe o 2º Valor: ");
+            valorB = Int32.Parse(Console.ReadLine());
 
-            do {
-                Console.WriteLine("DIgite: 1 para Adição");
+            do
+            {
+                Console.WriteLine("Digite: 1 para Adição");
                 Console.WriteLine("        2 para Subtração");
                 Console.WriteLine("        3 para Multiplicação");
                 Console.WriteLine("        4 para Divisão ");
                 menu = Int32.Parse(Console.ReadLine());
-                switch (menu) 
-                    {
+                switch (menu)
+                {
                     case 1:
                         adicao = (valorA + valorB);
                         Console.WriteLine("\n" + valorA + " + " + valorB + " = " + adicao);
@@ -47,7 +48,7 @@ namespace ExerciciosDois
                         divisao = (valorA / valorB);
                         Console.WriteLine(+valorA + " / " + valorB + " = " + divisao);
                         break;
-                    }
+                }
             } while (menu < 5);
         }
     }
