+++ 1MediaAlunos.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace ExerciciosDois
+namespace ListaExercicios2
 {
-    public class MediaAlunos 
+    public class MediaAlunos
     {
         public static void Media()
         {
@@ -12,10 +12,8 @@ namespace ExerciciosDois
             int nota3;
             int media1;
             // 1ª Forma Inserção de Dados:
-            Console.Write("Digite a 1ª Nota: ");
+            Console.WriteLine("Digite a 1ª Nota: ");
             nota1 = Int32.Parse(Console.ReadLine());
-            // 2ª Forma Inserção de Dados:
-            //nota2 = Convert.ToInt32(Console.ReadLine());
 
             Console.WriteLine("Digite a 2ª Nota: ");
             nota2 = Int32.Parse(Console.ReadLine());
@@ -23,17 +21,25 @@ namespace ExerciciosDois
             Console.WriteLine("Digite a 3ª Nota: ");
             nota3 = Int32.Parse(Console.ReadLine());
 
+            // 2ª Forma Inserção de Dados:
+            //nota2 = Convert.ToInt32(Console.ReadLine());
+
             media1 = ((nota1 + nota2 + nota3) / 3);
 
-            if (media1 >= 7) { // Maior ou Igual ao Resultado 21
+            if (media1 >= 7)
+            { // Maior ou Igual ao Resultado 21
                 Console.WriteLine("=================");
                 Console.WriteLine(" ALUNO APROVADO!");
                 Console.WriteLine("=================");
-            } else if (media1 >= 5) { // Menor ou Igual ao Resultado 20
+            }
+            else if (media1 >= 5)
+            { // Menor ou Igual ao Resultado 20
                 Console.WriteLine("====================");
                 Console.WriteLine(" ALUNO RECUPERAÇÃO!");
                 Console.WriteLine("====================");
-            } else { // Caso Contrário
+            }
+            else
+            { // Caso Contrário
                 Console.WriteLine("==================");
                 Console.WriteLine(" ALUNO REPROVADO!");
                 Console.WriteLine("==================");
