--- CincoStringPalavra.cs
@@ -1,20 +1,0 @@
-using System;
-
-namespace ExerciciosDois
-{
-    public class StPalavra 
-    {
-        public static void StPalavraLetra()
-        {
-        
-        Console.WriteLine("\nInforme a Palavra: ");
-        string palavra5 = Console.ReadLine();
-        
-        for (int i = 0; i < palavra5.Length; i++)
-            {
-                Console.WriteLine(palavra5[i]);
-            }
-        }
-        
-    }
-}
