+++ ListaExercicio02.md
@@ -1,6 +1,7 @@
 # Exercícios de Revisão
 
-1) Crie um programa que receba três notas calcule a média e diga se foi aprovado (`(nota1 + nota2 + nota3 / 3)`) e (`média maior que 7 para ser aprovado`).
+1) Crie um programa que receba três notas calcule a média e diga se foi aprovado 
+   (`(nota1 + nota2 + nota3 / 3)`) e (`média maior que 7 para ser aprovado`).
 2) Crie um programa que receba um número e indique o mês que representa.
 3) Crie um programa que simule um jogo da forca, com a entrada da palavra oculta, a quantidade de tentativas para acerto e depois solicite a entrada da palavra. A cada erro deve imprimir a quantidade de tentativas restantes e o 'membro' que foi marcado. Ao final deve imprimir se acertou ou não a palavra e se foi 'enforcado'.
 4) Crie um programa que imprima a tabuada de 0 a 10.
@@ -8,7 +9,7 @@
 6) Crie um programa que de a soma de todos os números ímpares e múltiplos de 7 entre 1 e 500.
 7) Crie um programa que tenha como entrada as notas de um aluno. Somente se encerrará quando for inserida uma nota negativa. Ao final deverá informar a média das notas.
 8) Crie um programa que receba um valor e imprima o fatorial desse número. Ex.: 3! = 3 x 2 x 1 = 6.
-9) Crie um programa que receba a altera e o peso e calcule o IMC:
+9)  Crie um programa que receba a altera e o peso e calcule o IMC:
 
 IMC | Classificação
 --- | ------
