+++ README.md
@@ -1,1 +1,13 @@
 # Lista-de-Exercicios-2-CSharp
+
+1 >>> Média 3 Alunos          
+2 >>> Digito Mês              
+3 >>> Jogo da Forca - Palavra 
+4 >>> Tabuada - Select        
+5 >>> String - Letra/Letra    
+6 >>> Soma dos Setes          
+7 >>> Média Alunos Negativo   
+8 >>> Fatorial do Número      
+9 >>> Cáculo do IMC           
+10 >>> Operações Matemáticas  
+11 >>> DESAFIO do Troco
